preparing for making alignment algorithm
