+++ Makefile
@@ -4,8 +4,7 @@ CXXFLAGS := -Wall -Wextra -Wno-missing-field-initializers -Wno-unused-variable \
 		 -Iinclude -std=c++23 -O2
 LDFLAGS := -Llib -lraylib -lGL -lm -lpthread -ldl -lrt -lX11
 
-SOURCES := ./src/main.cpp ./src/reader.cpp ./src/imgtools/filters.cpp ./src/imgtools/align.cpp \
-		  ./src/imgtools/imgtools.cpp
+SOURCES := ./src/main.cpp ./src/reader.cpp ./src/imgtools/filters.cpp ./src/imgtools/imgtools.cpp
 OBJECTS := $(patsubst ./src/%.cpp, ./build/%.o, $(SOURCES))
 OUTPUT := ./build/lagosta
 BUILD_DIRS := $(sort $(dir $(OBJECTS)))
